+++ lib/model/source/media_store_source.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:math';
 
 import 'package:aves/model/covers.dart';
 import 'package:aves/model/entry/entry.dart';
@@ -57,7 +56,7 @@ class MediaStoreSource extends CollectionSource {
   Future<void> _loadEssentials() async {
     final stopwatch = Stopwatch()..start();
     state = SourceState.loading;
-    await metadataDb.init();
+    await localMediaDb.init();
     await vaults.init();
     await favourites.init();
     await covers.init();
@@ -67,8 +66,8 @@ class MediaStoreSource extends CollectionSource {
       if (currentTimeZoneOffset != catalogTimeZoneOffset) {
         // clear catalog metadata to get correct date/times when moving to a different time zone
         debugPrint('$runtimeType clear catalog metadata to get correct date/times');
-        await metadataDb.clearDates();
-        await metadataDb.clearCatalogMetadata();
+        await localMediaDb.clearDates();
+        await localMediaDb.clearCatalogMetadata();
         settings.catalogTimeZoneRawOffsetMillis = currentTimeZoneOffset;
       }
     }
@@ -92,13 +91,13 @@ class MediaStoreSource extends CollectionSource {
       final topIds = settings.topEntryIds?.toSet();
       if (topIds != null) {
         debugPrint('$runtimeType refresh ${stopwatch.elapsed} load ${topIds.length} top entries');
-        topEntries.addAll(await metadataDb.loadEntriesById(topIds));
+        topEntries.addAll(await localMediaDb.loadEntriesById(topIds));
         addEntries(topEntries);
       }
     }
 
     debugPrint('$runtimeType refresh ${stopwatch.elapsed} fetch known entries');
-    final knownEntries = await metadataDb.loadEntries(origin: EntryOrigins.mediaStoreContent, directory: directory);
+    final knownEntries = await localMediaDb.loadEntries(origin: EntryOrigins.mediaStoreContent, directory: directory);
     final knownLiveEntries = knownEntries.where((entry) => !entry.trashed).toSet();
 
     debugPrint('$runtimeType refresh ${stopwatch.elapsed} check obsolete entries');
@@ -146,7 +145,7 @@ class MediaStoreSource extends CollectionSource {
     // clean up obsolete entries
     if (removedEntries.isNotEmpty) {
       debugPrint('$runtimeType refresh ${stopwatch.elapsed} remove obsolete entries');
-      await metadataDb.removeIds(removedEntries.map((entry) => entry.id).toSet());
+      await localMediaDb.removeIds(removedEntries.map((entry) => entry.id).toSet());
     }
 
     // verify paths because some apps move files without updating their `last modified date`
@@ -159,46 +158,44 @@ class MediaStoreSource extends CollectionSource {
     });
 
     // items to add to the collection
-    final pendingNewEntries = <AvesEntry>{};
+    final newEntries = <AvesEntry>{};
 
     // recover untracked trash items
     debugPrint('$runtimeType refresh ${stopwatch.elapsed} recover untracked entries');
     if (directory == null) {
-      pendingNewEntries.addAll(await recoverUntrackedTrashItems());
+      newEntries.addAll(await recoverUntrackedTrashItems());
     }
 
     // fetch new & modified entries
     debugPrint('$runtimeType refresh ${stopwatch.elapsed} fetch new entries');
-    // refresh after the first 10 entries, then after 100 more, then every 1000 entries
-    var refreshCount = 10;
-    const refreshCountMax = 1000;
-    final allNewEntries = <AvesEntry>{};
-    void addPendingEntries() {
-      allNewEntries.addAll(pendingNewEntries);
-      addEntries(pendingNewEntries);
-      pendingNewEntries.clear();
-    }
-
     mediaStoreService.getEntries(_safeMode, knownDateByContentId, directory: directory).listen(
       (entry) {
         // when discovering modified entry with known content ID,
         // reuse known entry ID to overwrite it while preserving favourites, etc.
         final contentId = entry.contentId;
         final existingEntry = knownContentIds.contains(contentId) ? knownLiveEntries.firstWhereOrNull((entry) => entry.contentId == contentId) : null;
-        entry.id = existingEntry?.id ?? metadataDb.nextId;
+        entry.id = existingEntry?.id ?? localMediaDb.nextId;
 
-        pendingNewEntries.add(entry);
-        if (pendingNewEntries.length >= refreshCount) {
-          refreshCount = min(refreshCount * 10, refreshCountMax);
-          addPendingEntries();
-        }
+        newEntries.add(entry);
       },
       onDone: () async {
-        addPendingEntries();
-
-        if (allNewEntries.isNotEmpty) {
+        if (newEntries.isNotEmpty) {
           debugPrint('$runtimeType refresh ${stopwatch.elapsed} save new entries');
-          await metadataDb.saveEntries(allNewEntries);
+          await localMediaDb.insertEntries(newEntries);
+
+          // TODO TLAD find duplication cause
+          final duplicates = await localMediaDb.searchLiveDuplicates(EntryOrigins.mediaStoreContent, newEntries);
+          if (duplicates.isNotEmpty) {
+            unawaited(reportService.recordError(Exception('Loading entries yielded duplicates=${duplicates.join(', ')}'), StackTrace.current));
+            // post-error cleanup
+            await localMediaDb.removeIds(duplicates.map((v) => v.id).toSet());
+            for (final duplicate in duplicates) {
+              final duplicateId = duplicate.id;
+              newEntries.removeWhere((v) => duplicateId == v.id);
+            }
+          }
+
+          addEntries(newEntries);
 
           // new entries include existing entries with obsolete paths
           // so directories may be added, but also removed or simply have their content summary changed
@@ -224,7 +221,7 @@ class MediaStoreSource extends CollectionSource {
         notifyAlbumsChanged();
 
         debugPrint('$runtimeType refresh ${stopwatch.elapsed} done');
-        unawaited(reportService.log('Source refresh complete in ${stopwatch.elapsed.inSeconds}s for ${knownEntries.length} known, ${allNewEntries.length} new, ${removedEntries.length} removed'));
+        unawaited(reportService.log('Source refresh complete in ${stopwatch.elapsed.inSeconds}s for ${knownEntries.length} known, ${newEntries.length} new, ${removedEntries.length} removed'));
       },
       onError: (error) => debugPrint('$runtimeType stream error=$error'),
     );
@@ -242,7 +239,7 @@ class MediaStoreSource extends CollectionSource {
     state = SourceState.loading;
 
     debugPrint('$runtimeType refreshUris ${changedUris.length} uris');
-    final uriByContentId = Map.fromEntries(changedUris.map((uri) {
+    final changedUriByContentId = Map.fromEntries(changedUris.map((uri) {
       final pathSegments = Uri.parse(uri).pathSegments;
       // e.g. URI `content://media/` has no path segment
       if (pathSegments.isEmpty) return null;
@@ -253,16 +250,16 @@ class MediaStoreSource extends CollectionSource {
     }).whereNotNull());
 
     // clean up obsolete entries
-    final obsoleteContentIds = (await mediaStoreService.checkObsoleteContentIds(uriByContentId.keys.toList())).toSet();
-    final obsoleteUris = obsoleteContentIds.map((contentId) => uriByContentId[contentId]).whereNotNull().toSet();
+    final obsoleteContentIds = (await mediaStoreService.checkObsoleteContentIds(changedUriByContentId.keys.toList())).toSet();
+    final obsoleteUris = obsoleteContentIds.map((contentId) => changedUriByContentId[contentId]).whereNotNull().toSet();
     await removeEntries(obsoleteUris, includeTrash: false);
-    obsoleteContentIds.forEach(uriByContentId.remove);
+    obsoleteContentIds.forEach(changedUriByContentId.remove);
 
     // fetch new entries
     final tempUris = <String>{};
     final newEntries = <AvesEntry>{}, entriesToRefresh = <AvesEntry>{};
     final existingDirectories = <String>{};
-    for (final kv in uriByContentId.entries) {
+    for (final kv in changedUriByContentId.entries) {
       final contentId = kv.key;
       final uri = kv.value;
       final sourceEntry = await mediaFetchService.getEntry(uri, null);
@@ -276,7 +273,7 @@ class MediaStoreSource extends CollectionSource {
             if (existingEntry != null) {
               entriesToRefresh.add(existingEntry);
             } else {
-              sourceEntry.id = metadataDb.nextId;
+              sourceEntry.id = localMediaDb.nextId;
               newEntries.add(sourceEntry);
             }
             final existingDirectory = existingEntry?.directory;
@@ -303,8 +300,22 @@ class MediaStoreSource extends CollectionSource {
     state = SourceState.ready;
 
     if (newEntries.isNotEmpty) {
+      await localMediaDb.insertEntries(newEntries);
+
+      // TODO TLAD find duplication cause
+      final duplicates = await localMediaDb.searchLiveDuplicates(EntryOrigins.mediaStoreContent, newEntries);
+      if (duplicates.isNotEmpty) {
+        unawaited(reportService.recordError(Exception('Refreshing entries yielded duplicates=${duplicates.join(', ')}'), StackTrace.current));
+        // post-error cleanup
+        await localMediaDb.removeIds(duplicates.map((v) => v.id).toSet());
+        for (final duplicate in duplicates) {
+          final duplicateId = duplicate.id;
+          newEntries.removeWhere((v) => duplicateId == v.id);
+          tempUris.add(duplicate.uri);
+        }
+      }
+
       addEntries(newEntries);
-      await metadataDb.saveEntries(newEntries);
       await analyze(analysisController, entries: newEntries);
     }
 
@@ -346,7 +357,7 @@ class MediaStoreSource extends CollectionSource {
   // vault
 
   Future<void> _loadVaultEntries(String? directory) async {
-    addEntries(await metadataDb.loadEntries(origin: EntryOrigins.vault, directory: directory));
+    addEntries(await localMediaDb.loadEntries(origin: EntryOrigins.vault, directory: directory));
   }
 
   Future<void> _refreshVaultEntries({
@@ -367,7 +378,7 @@ class MediaStoreSource extends CollectionSource {
         final sourceEntry = await mediaFetchService.getEntry(uri, null, allowUnsized: true);
         if (sourceEntry != null) {
           newEntries.add(sourceEntry.copyWith(
-            id: metadataDb.nextId,
+            id: localMediaDb.nextId,
             origin: EntryOrigins.vault,
           ));
         }