+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+## <a id="v1.11.10"></a>[v1.11.10] - 2024-09-01
+
+### Added
+
+- Swedish translation (thanks Shift18, Andreas Håll)
+
+### Changed
+
+- request notification permission when launching scanning service
+- upgraded Flutter to stable v3.24.1
+
+### Fixed
+
+- duplicates from new item loading/refreshing
+
 ## <a id="v1.11.9"></a>[v1.11.9] - 2024-08-07
 
 ### Added